analytics-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aio-pika
py-eureka-client
requests